+++ setup.cfg
@@ -64,6 +64,7 @@ niftypet=niftypet>=0.0.1
 console_scripts=
     amypet=amypet.cli:main
     amypet.gui=amypet.gui:main
+    amypet.web=amypet.web:main
 [options.packages.find]
 exclude=tests
 [options.package_data]
